fix network output for python 3
In python2 sockets require all data sent to be of type str. In python3 sockets require all data to be sent as bytes instead, in keeping with the great str shakeup of the 2->3 transition. Thus the network output must be retooled to coerce all inputs to bytes instead of str