Changed structure for NFA
nfa['transitions'] previously was a dictionary with a key of (state,
character) and a value of a list of destination states.  Changed it to
be a double dictionary where the nfa['transitions'] is a dictionary of
state to dictionary and the second dictionary was a dictionary of
transitions to characters.

This change was to make support for character class transitions easier